backend/requirements.txt
Flask==3.0.3
pillow==10.4.0
requests==2.32.3
transformers==4.44.1
torch==2.4.0
tensorflow==2.17.0
gunicorn==23.0.0